9th commit

+++ app/database.py
@@ -11,12 +11,12 @@ conn.autocommit = True  # не требует каждый раз вызыват
 
 class User:  # методы работы с таблицей users
     @staticmethod
-    def create_user(login, password, first_name, middle_name, sur_name, email, phone_number, pers_photo_data):
+    def create_user(login, password, first_name, middle_name, sur_name, email, phone_number, persphotodata):
         cur = conn.cursor()
         cur.execute("""
-            INSERT INTO users (login, password, firstName, middleName, surName, privileged, email, phoneNumber, persPhotoData)
+            INSERT INTO users (login, password, firstName, middleName, surName, privileged, email, phoneNumber, persphotodata)
             VALUES (%s, %s, %s, %s, %s, FALSE, %s, %s, %s)
-        """, (login, password, first_name, middle_name, sur_name, email, phone_number, pers_photo_data))
+        """, (login, password, first_name, middle_name, sur_name, email, phone_number, persphotodata))
 
     @staticmethod
     def find_by_login(login):
@@ -25,29 +25,32 @@ class User:  # методы работы с таблицей users
         user = cur.fetchone()
         return user
 
+    @staticmethod
+    def update_user():
+        pass
+
 
 class Post:  # методы работы с таблицей posts
     @staticmethod
-    def create_post(unique_id, user_login, title, content, tags, image_data):
+    def create_post(unique_id, owner_login, title, content, tags, image_data):
         cur = conn.cursor()
-        cur.execute("""
-            INSERT INTO posts (unique_id, user_login, title, content, tags, createdAt, imageData, viewCount, likesCount, dislikesCount)
+        cur.execute("""INSERT INTO posts (unique_id, owner_login, title, content, tags, createdAt, imageData, viewCount, likesCount, dislikesCount)
             VALUES (%s, %s, %s, %s, %s, NOW(), %s, 0, 0, 0)
-        """, (unique_id, user_login, title, content, tags, image_data))
+""", (unique_id, owner_login, title, content, tags, image_data))
 
     @staticmethod
     def get_all_posts(order='desc', page=1, limit=0, search=None):
         offset = (page - 1) * limit  # смещение пагинации
 
-        query = f'SELECT * FROM posts '
+        query = f"SELECT * FROM posts "
 
         if search:
             query += f"WHERE content LIKE '%{search}%' or title LIKE '%{search}%' or tags LIKE '%{search}%' "
 
-        query += f'ORDER BY createdAt {order} '
+        query += f"ORDER BY createdAt {order} "
 
         if limit != 0:
-            query += f'LIMIT {limit} OFFSET {offset};'
+            query += f"LIMIT {limit} OFFSET {offset};"
 
         cur = conn.cursor()
         cur.execute(query)
@@ -55,25 +58,41 @@ class Post:  # методы работы с таблицей posts
         return posts
 
     @staticmethod
-    def get_post_by_id(post_id):
+    def get_post_by_id(unique_id):
         cur = conn.cursor()
-        cur.execute("SELECT * FROM posts WHERE unique_id = %s;", (post_id,))
+        cur.execute("SELECT * FROM posts WHERE unique_id = %s;", (unique_id,))
         post = cur.fetchone()
         return post
 
     @staticmethod
-    def update_post(post_id, title, content, tags, image_data):
+    def update_post(unique_id, owner_login, **field):
         cur = conn.cursor()
-        cur.execute("""
-            UPDATE posts
-            SET title = %s, content = %s, tags = %s, imageData = %s
-            WHERE unique_id = %s;
-        """, (title, content, tags, image_data, post_id))
+
+        # Сначала проверяем, что логин пользователя совпадает с логином создателя поста
+        cur.execute("SELECT owner_login FROM posts WHERE unique_id = %s", (unique_id,))
+        result = cur.fetchone()
+
+        if result and result[0] == owner_login:
+            # Если логины совпадают, обновляем пост
+            fields = [f"{key} = %s" for key in field if field[key] is not None]
+            values = [field[key] for key in field if field[key] is not None]
+
+            if fields:
+                values.append(unique_id)
+                query = f"UPDATE posts SET {', '.join(fields)} WHERE unique_id = %s;"
+                cur.execute(query, values)
+                conn.commit()
 
     @staticmethod
-    def delete_post(post_id):
+    def delete_post(unique_id, owner_login):
         cur = conn.cursor()
-        cur.execute("DELETE FROM posts WHERE unique_id = %s;", (post_id,))
+
+        # Сначала проверяем, что логин пользователя совпадает с логином создателя поста
+        cur.execute("SELECT owner_login FROM posts WHERE unique_id = %s", (unique_id,))
+        result = cur.fetchone()
+
+        if result and result[0] == owner_login:
+            cur.execute("DELETE FROM posts WHERE unique_id = %s;", (unique_id,))
 
     @staticmethod
     def increment_view(post_id):
@@ -93,13 +112,20 @@ class Post:  # методы работы с таблицей posts
 
 class Comment:  # методы работы с таблицей comments
     @staticmethod
-    def create_comment(user_login, post_id, content, image_data, tags):
+    def create_comment(unique_id, owner_login, post_id, content):
         cur = conn.cursor()
-        cur.execute("""
-            INSERT INTO comments (user_login, post_id, content, imageData, tags, createdAt)
-            VALUES (%s, %s, %s, %s, %s, NOW())
-        """, (user_login, post_id, content, image_data, tags))
-        comment_id = cur.fetchone()[0]
+
+        # Проверяем, существует ли пост с указанным post_id
+        cur.execute("SELECT 1 FROM posts WHERE unique_id = %s", (post_id,))
+        post_exists = cur.fetchone()
+
+        # Создаем комментарий, если пост существует
+        if post_exists:
+            cur.execute("""
+                INSERT INTO comments (unique_id, owner_login, post_id, content, createdAt)
+                VALUES (%s, %s, %s, %s, NOW())
+            """, (unique_id, owner_login, post_id, content))
+            conn.commit()
 
     @staticmethod
     def get_comments_by_post(post_id, sort='date', order='desc', page=1, limit=10):
@@ -133,15 +159,23 @@ class Comment:  # методы работы с таблицей comments
         return comment
 
     @staticmethod
-    def update_comment(comment_id, content, image_data, tags):
+    def update_comment(comment_id, owner_login, content):
         cur = conn.cursor()
-        cur.execute("""
-            UPDATE comments
-            SET content = %s, imageData = %s, tags = %s
-            WHERE id = %s;
-        """, (content, image_data, tags, comment_id))
+        cur.execute("SELECT owner_login FROM posts WHERE unique_id = %s", (comment_id,))
+        result = cur.fetchone()
+
+        if result and result[0] == owner_login:
+            cur.execute("""
+                UPDATE comments
+                SET content = %s
+                WHERE id = %s;
+            """, (content, comment_id))
 
     @staticmethod
-    def delete_comment(comment_id):
-        cur = conn.cursor()
-        cur.execute("DELETE FROM comments WHERE id = %s;", (comment_id,))
+    def delete_comment(comment_id, owner_login):
+        cur = conn.cursor("SELECT owner_login FROM posts WHERE unique_id = %s", (comment_id,))
+        result = cur.fetchone()
+
+        if result and result[0] == owner_login:
+            cur.execute("")
+            cur.execute("DELETE FROM comments WHERE id = %s;", (comment_id,))
